Unified processing of resource classes and ordinary classes

+++ projects/valkyrie-ast/src/expression_level/dispatch.rs
@@ -60,7 +60,7 @@ impl PrettyPrint for PostfixCallPart {
 }
 
 impl ValkyrieNode for ExpressionKind {
-    fn get_range(&self) -> Range<usize> {
+    fn get_range(&self) -> Range<u32> {
         match self {
             Self::Placeholder => unreachable!(),
             Self::Null(node) => node.get_range(),
